Intelligent-Study-Assistant/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-common==1.1.28
azure-core==1.38.3
azure-identity==1.25.3
azure-search-documents==11.6.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.135.1
fastapi-mcp==0.4.0
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langsmith==0.7.17
markdown-it-py==4.0.0
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
oauthlib==3.3.1
openai==2.28.0
orjson==3.11.7
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
PyPDF2==3.0.1
pytest==9.0.3
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tomli==2.4.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
